+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the `vscode-cvs` extension will be documented in this fil
 
 ## Release Notes
 
+### 0.0.4
+
+Merge 'CodeCVS: add opened file in current window' and 'CodeCVS: add local directory to repository' commands
+'CodeCVS: compare': revert FS changes if CVS operation did not complete correctly
+
 ### 0.0.3
 
 Added new command: 'CodeCVS: compare'